% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{knit_to}
\alias{knit_to}
\title{Render a document to a specified directory}
\usage{
knit_to(output_dir)
}
\arguments{
\item{output_dir}{passed to [rmarkdown::render]}
}
\description{
Render a document to a specified directory
}
