% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{uconn_beamer}
\alias{uconn_beamer}
\title{uconn presentation engine}
\usage{
uconn_beamer(
  template = uconn_beamer_template(),
  toc = TRUE,
  theme = "CambridgeUS",
  includes = list()
)
}
\arguments{
\item{template}{see [rmarkdown::beamer_document]}

\item{toc}{not used}

\item{theme}{see [rmarkdown::beamer_document]}

\item{includes}{see [rmarkdown::beamer_document]}

\item{...}{passed to [rmarkdown::beamer_document]}
}
\description{
Build a beamer presentation with uconn branding.
}
