% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{uconn_document}
\alias{uconn_document}
\title{UConn document engine}
\usage{
uconn_document(
  ...,
  fig_width = 6,
  fig_height = 4,
  keep_tex = TRUE,
  toc = TRUE,
  number_sections = TRUE,
  includes = uconn_include()
)
}
\arguments{
\item{\dots}{passed to [rmarkdown::pdf_document]}

\item{fig_width}{see [rmarkdown::pdf_document]}

\item{fig_height}{see [rmarkdown::pdf_document]}

\item{keep_tex}{see [rmarkdown::pdf_document]}

\item{toc}{see [rmarkdown::pdf_document]}

\item{number_sections}{see [rmarkdown::pdf_document]}

\item{includes}{see [rmarkdown::pdf_document]}
}
\description{
UConn document engine
}
