% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{uconn_include}
\alias{uconn_include}
\title{Return a list of includes}
\usage{
uconn_include(
  header = system.file("rmarkdown/templates/uconn_document/resources/header.tex",
    package = "scsuconn"),
  ...
)
}
\arguments{
\item{header}{by default `header.tex`}

\item{...}{passed to [rmarkdown::includes]}
}
\description{
Return a list of includes
}
