% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{uconn_presentation}
\alias{uconn_presentation}
\title{uconn presentation engine}
\usage{
uconn_presentation(...)
}
\arguments{
\item{\dots}{passed to [rmarkdown::beamer_presentation]}
}
\description{
Build a beamer presentation with uconn branding.
}
